Factor out travis-haskell-env; install deps when needed.
It's too much of a pain to have the entire travis build script be in
"eval", so I've factored that part out, meaning we need 2 lines instead
of 1 in each repo's "script" section (1 for executing the script, 1 for
eval env).

We now build libsodium and c-toxcore if needed by the package.